requirements.txt
numpy
scipy
matplotlib
h5py
Pillow